enhanced login/signup pages with brand header and unified light green backgorund while moving the login and signup page code inside the auth folder.
